+++ CHANGELOG.md
@@ -6,9 +6,11 @@
     * add support connection using ssh tunnel to paramiko_shell
     * add rsync client authentication with ssh key
     * add management of redis cluster with sentinel
+    * add tunnel configuration for sql manager
 * Fixed ...
     * fixed parse_redis_uri to parse redis sentinel uri
     * replaced caracther | with + in simple.truncate
+    * renamed schema in db in SqlManager
 * Integrated ...
 * Various bugfixes
 * Removed